Added a mode to the DrawPathTool.
Wherever you click once it will go in "Point to Point Mode" and only add points when you press the mouse button. When you press the mouse button twice on the same point it will exit the DrawPathTool.

When you move the pointer while the mouse button is still pressed, it will create a path following the pointer. Releasing the mouse button will exit the DrawPathTool.